backend/services/chart_canslim_service/requirements.txt
yfinance>=0.2.36
mplfinance>=0.12.9b7
google-genai>=1.0.0
pandas>=2.0.0
pillow>=10.0.0
matplotlib>=3.8.0
python-dotenv>=1.0.0

# FastAPI 서버 (api.py)
fastapi>=0.100.0
uvicorn>=0.23.0
